moved SourcesEditor to use props and exporting component
