+++ .changeset/calm-feet-attack.md
@@ -1,0 +1,6 @@
+---
+"@inkbeard/budget-it": minor
+---
+
+- replaced SourcesEditor store with props.
+- exporting SourcesEditor component
