+++ .changeset/healthy-dolls-explode.md
@@ -1,0 +1,5 @@
+---
+"inkbeard": minor
+---
+
+- replaced local logic with SourcesEditor component
