+++ packages/budget-it/index.ts
@@ -2,10 +2,12 @@ import AddCategory from './src/components/AddCategory.vue';
 import SourceListing from './src/components/SourceListing.vue';
 import ExpenseItem from './src/components/ExpenseItem.vue';
 import ExpenseCategory from './src/components/ExpenseCategory.vue';
+import SourcesEditor from './src/components/SourcesEditor.vue';
 
 export {
   AddCategory,
   SourceListing,
   ExpenseItem,
   ExpenseCategory,
+  SourcesEditor,
 };
