+++ packages/budget-it/src/components/SourcesEditor.vue
@@ -1,25 +1,56 @@
 <script setup lang="ts">
   import { computed, ref } from 'vue';
   import { AppButton } from '@inkbeard/ui-vue';
-  import { useSourcesStore } from '@/stores/sources';
-  import SourceListing from '@/components/SourceListing.vue';
+  import SourceListing from './SourceListing.vue';
+  import type { ExpenseList, SourceList } from '../types';
 
-  const {
-    addSource,
-    deleteSource,
-    sourceList,
-    sourcesWithExpenses,
-  } = useSourcesStore();
+  const props = defineProps<{
+    expenseList: ExpenseList;
+  }>();
+  const defaultSourceId = defineModel<number | null>('defaultSourceId');
+  const sourceList = defineModel<SourceList>('sourceList', { required: true });
   /**
-   * Get and set the store's default source id.
+   * Get an alphabatize list of sources and their IDs.
    */
-  const defaultSourceId = computed({
-    get: () => useSourcesStore().defaultSourceId,
-    set: (sourceId) => {
-      useSourcesStore().defaultSourceId = sourceId;
-    },
-  });
+  const alphabaticSourceList = computed(() => Object.entries(sourceList.value)
+    .map(([id, source]) => ({ source, id: +id }))
+    .sort((a, b) => a.source.toLowerCase().localeCompare(b.source.toLowerCase())));
+  /**
+   * Get the expenses associated with the sources.
+   */
+  const sourcesWithExpenses = computed(() => Object.entries(props.expenseList)
+    .reduce((acc: any, [id, { sourceId }]) => {
+      if (acc[sourceId]) {
+        acc[sourceId].push(id);
+      } else {
+        acc[sourceId] = [id];
+      }
+
+      return acc;
+    }, {}));
   const isAdding = ref(false);
+  /**
+   * Add a new source to the current list of sources with the ID + 1 of the highest ID so far.
+   */
+  const addSource = (
+    {
+      sourceName,
+      isDefault,
+    }: {
+      sourceName: string,
+      isDefault: boolean
+    },
+  ) => {
+    const newSourceId = Math.max(...Object.keys(sourceList.value).map(Number)) + 1;
+
+    sourceList.value[newSourceId] = sourceName;
+
+    if (isDefault) {
+      defaultSourceId.value = newSourceId;
+    }
+
+    return newSourceId;
+  };
 </script>
 
 <template>
@@ -40,18 +71,17 @@
       v-model:default-source-id="defaultSourceId"
       v-model:is-editing="isAdding"
       :source-list="sourceList"
-      :sources-with-expenses="sourcesWithExpenses"
       @add-source="addSource"
     />
     <SourceListing
-      v-for="({ id }) in useSourcesStore().alphabaticSourceList"
+      v-for="({ id }) in alphabaticSourceList"
       :key="`source-list-${id}`"
       v-model:default-source-id="defaultSourceId"
       v-model:source-list="sourceList"
       :expenses-from-sources="sourcesWithExpenses[id]?.length"
       :source-id="id"
       :sources-with-expenses="sourcesWithExpenses"
-      @delete-source="deleteSource"
+      @delete-source="delete sourceList[id]"
     />
   </ul>
 </template>
